requirements.txt
# local package
-e .

pandas
pydicom
openpyxl
matplotlib
tqdm
pypng
python-opencv
sklearn
pandas-profiling[notebook]

# external requirements
click
Sphinx
coverage
awscli
flake8
python-dotenv>=0.5.1
